# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: TelPro - FileMaker App (Hive)
# during dev, should point to your local machine
#host: 52.34.207.5:5058
host: localhost:5000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getAllUsers:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: Get All Users
      summary: Get All Users
      # used as the method name of the controller
      operationId: getAllUsers
      parameters:
        - name: body
          in: body
          description: Get All Users
          schema:
           $ref: '#/definitions/getAllUsers'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getUserById:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: Get User Details
      summary: Get User Details
      # used as the method name of the controller
      operationId: getUserById
      parameters:
        - name: body
          in: body
          description: Get User Details
          schema:
           $ref: '#/definitions/getUserById'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getRole:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: get roles for user
      summary: get roles for user
      # used as the method name of the controller
      operationId: getRole
      parameters:
        - name: body
          in: body
          description: get roles for user
          schema:
           $ref: '#/definitions/getRole'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updateUserDetail:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: associate role  to user
      summary: associate role  to user
      # used as the method name of the controller
      operationId: updateUserDetail
      parameters:
        - name: body
          in: body
          description: associte role  to user
          schema:
           $ref: '#/definitions/updateUserDetail'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/removeUserPic:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: remove user pic
      summary: remove User Pic
      # used as the method name of the controller
      operationId: removeUserPic
      parameters:
        - name: body
          in: body
          description: remove User Pic
          schema:
           $ref: '#/definitions/removeUserPic'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/updateUserPic:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: Update User Profile Pic
      summary: Update User Profile Pic
      # used as the method name of the controller
      operationId: updateUserPic
      consumes:
        - multipart/form-data
        - application/x-www-form-urlencoded
        - binary
      parameters:
        - in: formData
          name: id
          description: Contact Id
          type: string
          required: true                
        - in: formData
          name: file
          description: Update user Profile Pic
          type: file
          required: true                      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/checkToken:
    # binds a127 app logic to a route
    x-swagger-router-controller: users
    post:
      tags: ['User Management']
      description: remove user pic
      summary: checkToken
      # used as the method name of the controller
      operationId: checkToken
      parameters:
        - name: body
          in: body
          description: checkToken
          schema:
           $ref: '#/definitions/checkToken'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  getAllUsers:
    type: object
    properties:
      companyId:
        type: string 
  getUserById:
    type: object
    properties:
      companyUserId:
        type: string
...
  getRole:
    type: object
    properties:
      companyId:
        type: string
  updateUserDetail:
    type: object
    properties:
      userId:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
      firstname:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
      lastname:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
      name:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
  removeUserPic:
    type: object
    properties:
      userId:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             
      path:
        type: string                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          
...
  checkToken:
    type: object
    properties:
      bearerToken:
        type: string  
...